// cli/src/_modules/elevenlabs/_enums/el-languages.enum.ts
export enum EL_Languages {
  // Major languages
  english = 'en',
  spanish = 'es',
  french = 'fr',
  german = 'de',
  italian = 'it',
  portuguese = 'pt',
  russian = 'ru',
  japanese = 'ja',
  korean = 'ko',
  chinese = 'zh',
  
  // European languages
  dutch = 'nl',
  swedish = 'sv',
  norwegian = 'no',
  danish = 'da',
  finnish = 'fi',
  polish = 'pl',
  czech = 'cs',
  hungarian = 'hu',
  romanian = 'ro',
  bulgarian = 'bg',
  greek = 'el',
  turkish = 'tr',
  
  // Asian languages
  hindi = 'hi',
  thai = 'th',
  vietnamese = 'vi',
  indonesian = 'id',
  malay = 'ms',
  filipino = 'tl',
  
  // Middle Eastern languages
  arabic = 'ar',
  hebrew = 'he',
  persian = 'fa',
  
  // African languages
  swahili = 'sw',
  zulu = 'zu',
  afrikaans = 'af',
  
  // Other languages
  catalan = 'ca',
  basque = 'eu',
  galician = 'gl',
  icelandic = 'is',
  latvian = 'lv',
  lithuanian = 'lt',
  estonian = 'et',
  slovak = 'sk',
  slovenian = 'sl',
  croatian = 'hr',
  serbian = 'sr',
  macedonian = 'mk',
  albanian = 'sq',
  ukrainian = 'uk',
  belarusian = 'be',
  georgian = 'ka',
  armenian = 'hy',
  azerbaijani = 'az',
  kazakh = 'kk',
  kyrgyz = 'ky',
  uzbek = 'uz',
  tajik = 'tg',
  turkmen = 'tk',
  mongolian = 'mn',
  nepali = 'ne',
  sinhala = 'si',
  bengali = 'bn',
  urdu = 'ur',
  punjabi = 'pa',
  gujarati = 'gu',
  marathi = 'mr',
  kannada = 'kn',
  telugu = 'te',
  malayalam = 'ml',
  tamil = 'ta',
  odia = 'or',
  assamese = 'as',
  manipuri = 'mni',
  konkani = 'kok',
  sanskrit = 'sa',
  tibetan = 'bo',
  dzongkha = 'dz',
  burmese = 'my',
  lao = 'lo',
  khmer = 'km',
  maldivian = 'dv',
  mongolianCyrillic = 'mn-Cyrl',
  mongolianTraditional = 'mn-Mong',
  uyghur = 'ug',
  tatar = 'tt',
  bashkir = 'ba',
  chuvash = 'cv',
  mordvin = 'mdf',
  udmurt = 'udm',
  komi = 'kv',
  mari = 'chm',
  erzya = 'myv',
  moksha = 'mdf',
  karelian = 'krl',
  veps = 'vep',
  livvi = 'olo',
  ludic = 'lud',
  votic = 'vot',
  ingrian = 'izh',
  livonian = 'liv',
  selkup = 'sel',
  nenets = 'yrk',
  nganasan = 'nio',
  enets = 'enf',
  dolgan = 'dlg',
  evenki = 'evn',
  even = 'eve',
  nanai = 'gld',
  udege = 'ude',
  oroch = 'oac',
  orok = 'oaa',
  nivkh = 'niv',
  ainu = 'ain',
  ryukyuan = 'ryu',
  japaneseOld = 'ja-old',
  koreanOld = 'ko-old',
  chineseTraditional = 'zh-Hant',
  chineseSimplified = 'zh-Hans',
  chineseCantonese = 'zh-yue',
  chineseHakka = 'zh-hak',
  chineseMinNan = 'zh-nan',
  chineseWu = 'zh-wuu',
  chineseXiang = 'zh-hsn',
  chineseGan = 'zh-gan',
  chineseJin = 'zh-cjy',
  chineseHuizhou = 'zh-hui',
  chinesePinghua = 'zh-png',
  chineseDungan = 'zh-dng',
  chineseXibe = 'zh-sjo',
  chineseEwenki = 'zh-ewe',
  chineseOroqen = 'zh-orh',
  chineseHezhen = 'zh-hez',
  chineseDaur = 'zh-dta',
  chineseBonan = 'zh-peh',
  chineseDongxiang = 'zh-sce',
  chineseTu = 'zh-mjg',
  chineseSalar = 'zh-slr',
  chineseYugur = 'zh-ybe',
  chineseBaoan = 'zh-bpe',
  chineseKangjia = 'zh-kxs',
  chineseSanta = 'zh-bft',
  chineseMonguor = 'zh-mgr',
  chineseDong = 'zh-kmr',
  chineseBouyei = 'zh-pcc',
  chineseDai = 'zh-tai',
  chineseZhuang = 'zh-zha',
  chineseMulao = 'zh-mlm',
  chineseMaonan = 'zh-mmm',
  chineseGelao = 'zh-gio',
  chineseLahu = 'zh-lhu',
  chineseJingpo = 'zh-kac',
  chineseLisu = 'zh-lis',
  chineseNaxi = 'zh-nbf',
  chineseBai = 'zh-bca',
  chinesePumi = 'zh-pum',
  chineseNu = 'zh-nuf',
  chineseAchang = 'zh-acn',
  chineseDeang = 'zh-dng',
  chineseJing = 'zh-jng',
  chineseTujia = 'zh-tji',
  chineseShe = 'zh-shx',
  chineseGaoshan = 'zh-ais',
  chineseLhoba = 'zh-lbo',
  chineseMonba = 'zh-mnp',
  chineseTatar = 'zh-ttr',
  chineseUzbek = 'zh-uzb',
  chineseKazakh = 'zh-kaz',
  chineseKyrgyz = 'zh-kir',
  chineseRussian = 'zh-rus',
  chineseKorean = 'zh-kor',
  chineseMongolian = 'zh-mon',
  chineseTibetan = 'zh-bod',
  chineseUyghur = 'zh-uig',
  chineseYi = 'zh-iii',
  chineseMiao = 'zh-hmd'
}